refactor(node): canvas routes Phase 2 — presence, state, flow-score, mood (task-1773689755389)
- Extracted GET /canvas/presence, /canvas/state, /canvas/flow-score, /canvas/team/mood
- Now 8 canvas routes in plugin (up from 4 in Phase 1)
- server.ts reduced by ~250 lines
- Deps: canvasStateMap, agentIdentityColors, getDb, flowExpressionLog
- formatRecency exported as shared helper
- Route-docs: 571/571, tests: 2455 pass, 0 regressions
